Add results generation.
Clip negative coordindates AABB in json to 0 to prevent crash while accessing the image in LBP.
Generate some results with pictures.
Add generation result function that take a random fish ( FishName, that mean we test all picture of a fish in a campaign ), and several others fishes and process the distance between LBP then write the results in a csv file with the subject FishName as name.